wip: proximity animation

+++ src/Light/AnimationConfig.h
@@ -5,9 +5,9 @@
 #include "AnimateFunc.h"
 
 struct RgbColor {
-  int red;
-  int green;
-  int blue;
+  byte red;
+  byte green;
+  byte blue;
 };
 
 struct AnimationConfig {
